+++ example/dsa_kit_example.dart
@@ -6,6 +6,8 @@ void main() {
   minHeap.push(5);
   minHeap.push(2);
   minHeap.push(8);
+  print(minHeap.printHeap());
+
   print('MinHeap peek: ${minHeap.peek()}'); // 2
   print('MinHeap pop: ${minHeap.pop()}'); // 2
   print('MinHeap pop: ${minHeap.pop()}'); // 5
@@ -15,6 +17,7 @@ void main() {
   maxHeap.push(5);
   maxHeap.push(2);
   maxHeap.push(8);
+  print(maxHeap.printHeap());
   print('MaxHeap peek: ${maxHeap.peek()}'); // 8
   print('MaxHeap pop: ${maxHeap.pop()}'); // 8
   print('MaxHeap pop: ${maxHeap.pop()}'); // 5