+++ lib/heap/heap_base.dart
@@ -2,8 +2,14 @@
 /// SOLID Principle: Interface Segregation + Dependency Inversion
 abstract class Heap<T extends Comparable> {
   void push(T value);
+
   T pop();
+
   T peek();
+
+  String printHeap();
+
   bool get isEmpty;
+
   int get length;
 }